+++ packages/moonstone/Slider/tests/Slider-specs.js
@@ -27,6 +27,31 @@ describe('Slider Specs', () => {
 		expect(actual).to.equal(expected);
 	});
 
+	it('should not fire change if clicked but value has not changed', function () {
+		const handleChange = sinon.spy();
+		const value = 50;
+
+		const slider = mount(
+			<Slider
+				min={0}
+				max={100}
+				value={value}
+				step={1}
+				onChange={handleChange}
+			/>
+		);
+
+		slider.find(`.${css.input}`).simulate('mousedown', {value});
+		slider.find(`.${css.input}`).simulate('change', {target: {value}});
+		// Technically it should be mousedown -> change -> mouseup, but simulating mouse down and mouse up on same element does seem to fire click event
+		// Simulating click event also doesn't fire mouse down or mouse up event, so we're okay.
+		slider.find(`.${css.input}`).simulate('click', {value});
+
+		const expected = 0;
+		const actual = handleChange.callCount;
+		expect(actual).to.equal(expected);
+	});
+
 	it('should not fire change event more than once', function () {
 		const handleChange = sinon.spy();
 		const value = 25;
@@ -41,7 +66,7 @@ describe('Slider Specs', () => {
 			/>
 		);
 
-		slider.find(`.${css.input}`).simulate('change', {target: {value}});
+		slider.find(`.${css.input}`).simulate('click', {value});
 
 		const expected = true;
 		const actual = handleChange.calledOnce;
@@ -64,6 +89,7 @@ describe('Slider Specs', () => {
 		);
 
 		slider.find(`.${css.input}`).simulate('change', {target: {value}});
+		slider.find(`.${css.input}`).simulate('click', {value});
 
 		const expected = value;
 		const actual = handleChange.args[0][0].value;